ip_defang
Defanging an IP address is a process of modifying the IP address so that it is not clickable or executable in a way that could potentially cause harm if clicked accidentally. Below is a simple Python script that defangs an IP address. This script can be run on Kali Linux or any other system with Python installed.